WIP Implementing HTTP server in JS the Bricks way.
* Split the modules into more files.
* Added the HTTP(port) call support.
* Added more tests.
