+++ lib/bricks-net-api.js
@@ -1,38 +1,28 @@
 'use strict';
 
-var when = require('when');
+exports.URL = require('./bricks-net-url').URL;
 
-exports.URL = require('./bricks-net-api-types').URL;
+exports.DefaultContentType = require('./bricks-net-http-request').DefaultContentType;
 
-exports.DefaultContentType = require('./bricks-net-api-types').DefaultContentType;
+exports.HTTPHeaders = require('./bricks-net-http-headers').HTTPHeaders;
 
-exports.HTTPHeaders = require('./bricks-net-api-types').HTTPHeaders;
-
-exports.HTTPResponse = require('./bricks-net-api-types').HTTPResponse;
+exports.HTTPResponse = require('./bricks-net-http-response').HTTPResponse;
 
 exports.HTTPResponseCode = require('./bricks-net-http-codes').HTTPResponseCode;
 exports.HTTPResponseCodeAsString = require('./bricks-net-http-codes').HTTPResponseCodeAsString;
 
-exports.GET = require('./bricks-net-api-types').GET;
-exports.POST = require('./bricks-net-api-types').POST;
-exports.POSTFromFile = require('./bricks-net-api-types').POSTFromFile;
+exports.GET = require('./bricks-net-http-request').GET;
+exports.POST = require('./bricks-net-http-request').POST;
+exports.POSTFromFile = require('./bricks-net-http-request').POSTFromFile;
 
-exports.KeepResponseInMemory = require('./bricks-net-api-types').KeepResponseInMemory;
-exports.SaveResponseToFile = require('./bricks-net-api-types').SaveResponseToFile;
+exports.KeepResponseInMemory = require('./bricks-net-http-response').KeepResponseInMemory;
+exports.SaveResponseToFile = require('./bricks-net-http-response').SaveResponseToFile;
 
-exports.HTTP = require('./bricks-net-api-types').HTTP;
+exports.HTTP = require('./bricks-net-http').HTTP;
 
-exports.JSON = function (arg) {
-	return when(arg).then(function (ret) {
-		return JSON.stringify(ret);
-	});
-};
+exports.JSON = require('./bricks-json').JSON;
+exports.JSONParse = require('./bricks-json').JSONParse;
 
-exports.JSONParse = function (arg) {
-	return when(arg).then(function (ret) {
-		return JSON.parse(ret);
-	});
-};
 
 /**
  * Creates a custom inspect function for the API items.
