+++ lib/bricks-net-exceptions.js
@@ -36,6 +36,7 @@ exports.HTTPException = makeCppException(exports.NetworkException, "HTTPExceptio
 exports.HTTPNoBodyProvidedException = makeCppException(exports.HTTPException, "HTTPNoBodyProvidedException");
 exports.HTTPRedirectNotAllowedException = makeCppException(exports.HTTPException, "HTTPRedirectNotAllowedException");
 exports.HTTPRedirectLoopException = makeCppException(exports.HTTPException, "HTTPRedirectLoopException");
+exports.CannotServeStaticFilesOfUnknownMIMEType = makeCppException(exports.HTTPException, "CannotServeStaticFilesOfUnknownMIMEType");
 
 // AttemptedToSendHTTPResponseMoreThanOnce is a user code exception; not really an HTTP one.
 exports.AttemptedToSendHTTPResponseMoreThanOnce = makeCppException(Exception, "AttemptedToSendHTTPResponseMoreThanOnce");
